Experiment with a gdb/gcc dejagnu test runner in the lab.
This is a bit cheesy just slapping it in as an option to the ClangBuilder, so
if anyone has ideas how they'd like to see the variations, extension points, 
etc, implemented - feel free to let me know in post-commit review.

I've done some experimentation with this locally, but without a mac box to run
it against, so I'll be testing this in the lab with incremental/experimental
commits for now.

llvm-svn: 161799