Remove the llvm-i686-debian builder.
It was failing to build for a long time, and when it came back it had a
failing test.

I was able to reproduce it on gcc15. It looks like llvm-mc is miscompiled by
gcc 4.4 in 32 bits mode. It crashes when running

llvm-mc -triple x86_64-pc-win32 -filetype=obj ~/llvm/test/MC/COFF/section-name-encoding.s

Since we are about to drop support for gcc 4.4, just drop the bot.

llvm-svn: 193728