Move sanitizer bots's configs from an external repo to zorg
Summary:
Currently, we store the configs for sanitizer bots in a separate repository:
https://code.google.com/p/address-sanitizer/source/browse/#svn%2Ftrunk%2Fbuild%2Fscripts%2Fslave
We are planing to add more hardware to the bots, and before that we would like to
perform some cleanup.
First (this change) we want to move the config files from the external repo to zorg.
Next, we'd like to change zorg/buildbot/builders/SanitizerBuilder.py to use the configs
from zorg instead of the extarnal repo.
Then we'll add some new bots.

Test Plan: n/a

Reviewers: eugenis, samsonov, gkistanova

Subscribers: rnk, rengolin, jfb, llvm-commits

Differential Revision: http://reviews.llvm.org/D5531

llvm-svn: 219136