[zorg] llgo: add logfiles to build_libgo
Summary:
llgo's "libgo" target builds libgo by executing
a Make subprocess; the Make build sends output
to log files on disk, and not to stdio. On the
llgo-builder slave, the libgo build step takes
more than the default timeout (1200s) to complete,
which is killed due to lack of output on stdio.

This diff adds the log files to the build step,
so the step doesn't time out and to assist in
diagnosing build failures. We also explicitly
specify the libgo and llgoi targets as steps.

Test Plan: Tested with docker-zorg.

Reviewers: pcc, gkistanova

Reviewed By: gkistanova

Subscribers: axw, llvm-commits

Differential Revision: http://reviews.llvm.org/D9283

llvm-svn: 235938