[zorg] Add support for uploading artifacts to the 'llvmlab bisect' bucket and enable this for clang-cmake-mips.
Summary:
Adds a getClangCMakeGCSBuildFactory() based on getClangCMakeBuildFactory(). This
function adds a new keyword argument to getClangCMakeBuildFactory():
    stage1_upload_directory - This is the bucket subdirectory to upload to and
                              should match the builder name.

The buildslave must have gsutil installed to use this argument and must have an
appropriate config installed. If the config is not installed in gsutil's search
path then the BOTO_CONFIG envvar can be used to specify it. Additionally the
BUCKET envvar can be used (as per 'llvmlab bisect') to specify an alternate GCS
bucket to use.

Reviewers: gkistanova

Subscribers: chatur01, llvm-commits, cmatthews, gkistanova

Differential Revision: http://reviews.llvm.org/D14866

llvm-svn: 254717