[zorg] Convert llvm-mips-linux builder to cmake.
Duplicated getLLVMBuildFactory and adapted it to cmake. The original should be
removed now that autoconf is not supported but other builders still reference
it.

Drop the following arguments from cmake version since nobody uses them anymore:
* llvmgccdir
* valgrind
* examples
* expensive_checks
* enable_targets
In the case of valgrind, 'llvm-x86_64-linux-vg_leak' referenced it but this
builder is disabled.

llvm-svn: 258949