update Xcode LLDB build to use the just-build debugserver
We are hitting issues where newer lldb code that requires
a newer debugserver is failing to launch debugserver
successfully due to using the older Xcode Apple-signed
debugserver.  A failure to launch debugserver during
the test suite run has dire consequences for the
host as it causes a bunch of test inferiors to be
launched but never disposed of properly.

This change forces CI to use an lldb_codesign key
as described in the lldb docs folder.

rdar://problem/27995417 tracks the "dire consequences"
issue.

llvm-svn: 279663