[buildbot] Add slave to test libomp for ppc64le targets.
Summary:
This patch adds a slave to test libomp for powerpc64 little endian. This configuration is known to work but didn't have a bot to actually do the testing. 

The configuration uses the existing builder infrastructure already in place for x86. The intent is to extend the testing to cover OpenMP offloading (libomptarget) when it is ready in the OpenMP project. The slave has an NVIDIA K40 GPU ready for that purpose.

Reviewers: hfinkel, ABataev, gkistanova, Hahnfeld

Subscribers: caomhin, grokos, arpith-jacob, carlo.bertolli, llvm-commits

Differential Revision: https://reviews.llvm.org/D28492

llvm-svn: 291567