[zorg] Remove clang-cmake-aarch64-42vma builder
... and re-purpose hardware for 2nd clang-cmake-aarch64-full builder.
Hardware hosting "aarch64-full" builder is not very stable, and we
want to use a known-stable box for this important builder.
Hardware hosting "aarch64-42vma" builder seems to be solid, and this
builder is no longer very valuable.
We are going to run "aarch64-full" builder on the 2 slaves in parallel
until unstable-ish box crashes again. At that point we are going
to dicommision that box and leave aarch64-full bot running on 1 slave.

Patch by Maxim Kuvyrkov.

Differential Revision: https://reviews.llvm.org/D37459

llvm-svn: 312647