Use check-all on the android buildbot
Build/test all of compiler-rt as a single step using check-all.
This reduces the number of steps on the buildbot page, and automatically picks up any new tests (like ubsan-minimal).

It would be great if the "unit" tests could also run through lit. This is out of scope of this change.

llvm-svn: 315521