[zorg] Update status reporting of reverse-iteration builder
Summary:
The mail notification to "interested users" has been disabled for the
reverse-iteration builder. This patch turns the mail notification on so that
interested users are notified of new failures on the reverse-iteration
builder. Notice that only transitions from pass to fail on the builder will
trigger mail notification. Subsequent failures will not be reported repeatedly.

Reviewers: mgrang, zinob, gkistanova, vleschuk, dblaikie

Reviewed By: dblaikie

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D39258

llvm-svn: 316603