Open source task tool + tasks
This open sources the 'task' tool which we created and used for internal
jobs for a while. It is basically a thin layer that allows you to put
self-contained task descriptions into a git repository. Those can then
be used as part of CI jobs, but also allow you to run tasks locally
without jenkins installed.

Also add a bunch of tasks and cmake caches to perfrom some lnt-ctmask,
lnt-test-suite and compiler-verifiers builds.

This will initially be used by the greendragon jobs.

llvm-svn: 317539