jenkins: Do pseudo checkouts in relay jobs
Relay jobs lacked blamelists/changelogs. This happened naturally because
we cannot communicate parameters/extra info to them to not loose the
feature of jenkins merging multiple jobs in a queue.

To get the changelists back the relay jobs are now modified to do a
pseudo checkout with all the relevant repositories (using depth: zero so
we shouldn't really spend time checking out files to disk) to get the
changelists back.

llvm-svn: 318242