Add a task for zorg tests
Note: We need FileCheck and not to run the tests. It would be silly to
checkout a complete llvm and build them just for this. So instead I am
doing the gross thing and put pre-built binaries into the repository
(they are small).

The binaries were created from r317641 like this:

$ mkdir sizebuild
$ cd sizebuild
$ cmake -GNinja -DCMAKE_BUILD_TYPE=MINSIZEREL -DLLVM_ENABLE_ASSERTIONS=OFF -DLLVM_ENABLE_BACKTRACES=OFF -DLLVM_ENABLE_CRASH_OVERRIDES=OFF -DLLVM_ENABLE_LTO=ON -DLLVM_ENABLE_LIBEDIT=OFF -DLLVM_ENABLE_THREAD=OFF -DLLVM_ENABLE_THREAD=OFF -DLLVM_ENABLE_TERMINFO=OFF -DLLVM_ENABLE_ZLIB=OFF -DLLVM_BUILD_TOOLS=OFF $LLVM_PATH
$ ninja bin/FileCheck bin/not
$ strip bin/FileCheck
$ strip bin/not

llvm-svn: 318253