[zorg] Add linaro-armv8-01-arm-* slaves
Summary:
We are going to migrate most of AArch32 builders from armv7
TK1 hardware to armv8 boxes.  Since the architecture is slightly
different (armv7 vs armv8) we shall take care to not connect both
armv7 and armv8 for the same builder simultaneously.
We plan to allocate freed up TK1 boards to performance and full bots.

Reviewers: gkistanova, rovka

Reviewed By: rovka

Subscribers: aemerson, kristof.beyls, llvm-commits

Differential Revision: https://reviews.llvm.org/D42040

llvm-svn: 323966