Replace clang-x86-windows-msvc2015 with an x64 VS 2017 build script
Summary:
Use the annotated buildbot steps that the Windows ThinLTO bot uses, and
make them find and prefer VS 2017 by default.

This will rename the builder to clang-x64-windows-msvc.

Once the master restarts, we'll be able to more test and make changes to
the buildbot script. I tested it locally, and it does a two-stage,
64-bit build with VS 2017 as the stage1 compiler.

Reviewers: inglorion, gkistanova

Subscribers: mehdi_amini, dexonsmith, llvm-commits

Differential Revision: https://reviews.llvm.org/D51746

llvm-svn: 342587