[Zorg] Explicitly use `-stdlib=libstdc++` for Android.
Summary:
This builder uses an old version of the NDK with a just-built version
of the Clang. This builder is making use of an NDK standalone
toolchain. For older NDKs, these toolchins have the C++ stdlib headers
and libraries installed to the GNU stdlib paths.

This causes an issue when I try to update the default -stdlib for
Android targets in Clang as this builder will begin looking in the
libc++ paths when it does not have an STL available at that path.

Explicitly use `-stdlib=libstdc++` to keep the old tools working with
a new Clang until this builder is using an NDK new enough to not need
it (NDK r19).

Reviewers: eugenis

Reviewed By: eugenis

Subscribers: srhines, llvm-commits

Tags: #zorg

Differential Revision: https://reviews.llvm.org/D53182

llvm-svn: 344370