[buildbot, windows] Update the way the visual studio environment is set
Summary: With vs2017 there were several changes to the installation paths for VS including the vcvarsall.bat script used for setting the environment. To correctly get the paths for the newer versions of VS, we should use vswhere instead. This is similar to how the annotated builder sets the environment as well

Reviewers: gkistanova

Reviewed By: gkistanova

Subscribers: asmith, llvm-commits

Differential Revision: https://reviews.llvm.org/D53158

llvm-svn: 345309