Reorganise Linaro Armv7 and Armv8 builders.
Due to limited amount of native Armv7 hardware, and the limited advantage
over running on Armv8 hardware, we are moving some of our v7 bots to run
on more powerful Armv8 hardware. We will also remove some of the v8 bots
that were identical configurations to the Armv7 besides targeting Armv8.

The remaining bots running on native v7 hardware make up 3 groups:
- 4x Tk1s running clang-cmake-thumbv7-full-sh
- 4x TK1s running clang-cmake-armv7-full
- 1x Tk1s running clang-native-arm-lnt-perf and libcxx bots

The clang-cmake-*-full* will do a 2-stage build and run the test-suite to
maximize the chance of catching a wrong instruction being generated.

Differential Revision: https://reviews.llvm.org/D66427

llvm-svn: 369372