Migrate the annotated buildbots to github
Summary:
Reuse the base LLVMBuildFactory logic for checking out and updating
llvm-project instead of implementing it in the script.

There are advantages to putting the 'git' logic in the script, but since
we need a master restart anyway to move to git, we might as well reuse
the buildbot git step for now. I'm not sure what the best practices are
for incrementally updating a git repository anyway.

Reviewers: inglorion, thakis

Subscribers: gkistanova, llvm-commits

Differential Revision: https://reviews.llvm.org/D69450