Conditionally build LLD for sanitizer bots.
7b38316441a50c2a3f48cd3e5691c85b7234085f made a change that
unconditionally started building LLD, as it's not present on the system
for x86_64 sanitizer buildbots. Unfortunately, the ppc64be bot doesn't
configure LLD as part of the build during the cmake step, and thus fails
when trying to build LLD.

Fix it so that we only conditionally build LLD, if $CHECK_LLD != 0.