Remove LLVM from depset for the libc++ builders.
This prevents every LLVM commit from triggering the libc++ buildbots,
which is a waste of resources because the only parts of LLVM that
libc++ depends on are the CMake files and LIT harness.

These file change infrequently enough that it seems OK to occasionally
fail to trigger a build when they change.