[zorg] Add extra argument passing to getAnnotatedBuildFactory.
Summary:
This allows the passing of extra arguments from builders to annotated script runners.
Originally I was planning on adding the ability to pass `extra_cmake_args`, however this approach seems more general to fit many needs.
For libc builders this will allow us to pass `--asan` and in the script this will add in the `extra_cmake_args`.

Reviewers: gkistanova

Reviewed By: gkistanova

Tags: #zorg

Differential Revision: https://reviews.llvm.org/D74193