clangd-ubuntu-tsan: Fix arguments passed to CMake
The builds are currently failing:
http://lab.llvm.org:8014/#/builders/136/builds/5

Presumably, the problem is that extra `"` make it into the values passed to
CMake somehow through the shell or invocation, and that causes unexpected
failures. The buildbot configurations around clangd-ubuntu-tsan never enclose
values in additional `"` so this patch also makes it more coherent with the
rest of existing code.

Differential Revision: https://reviews.llvm.org/D91566