Start Docker-izing sanitizer-windows
This defines a new docker image for sanitizer-windows based on the
existing docker image. I chose to dockerize sanitizer-windows first,
because it cycles quickly and doesn't do a two-stage build.

My plan is to have both the existing sanitizer-windows bot and the new
one run concurrently, but the new one will connect to the silent,
staging buildmaster (lab.llvm.org:9994). Once we are ready to cut over,
I will turn off the old sanitizer-windows VM and update this port in the
docker image and redeploy.

I am chosing to stick with the buildbot annotation script approach,
because it allows for much faster iteration: I can adjust paths for the
VS installation for example simply by committing to zorg. However, both
the live and staging bot will use the same script, so the same script
must work for both sanitizer-windows bots at the same time!

Differential Revision: https://reviews.llvm.org/D93340