[zorg] Support other relative scripts in AnnotatedBuilder
AnnotatedBuilder should be able to launch both python and bash
scripts. It should launch with python for scripts that specify
the script_interpreter as python, or allow the user to specify
another interpreter. Otherwise it will execute an empty
script_interpreter script directly. Also, move the --jobs option
to be the default of extra_args, otherwise allow the user to
override extra_args.

Reviewed By: tra, gkistanova

Differential Revision: https://reviews.llvm.org/D100666