Add SKIPPED to known result names.
A new test result on the buildbot workers, "SKIPPED", was introduced by D102754.

This was not specifically handled in LitTestCommand.py, so some build results
displayed, "Unexpected test result output SKIPPED", without the number of tests
producing that result.

This patch is to handle "SKIPPED" and display the number of tests producing that
result, as "X skipped unit tests".

The patch will not turn a red build into a green one, just display the relevant
number of tests.

I don't have a local copy of the buildbot infrastructure to test this for side
effects, but @gkistanova staged the change 2 months ago (June 2021), and it is
displaying the results as intended.

The first build of interest, showing "Unexpected test result output SKIPPED",
was:
https://lab.llvm.org/staging/#/builders/145/builds/996

A recent build showing the intended "X skipped unit tests" is:
https://lab.llvm.org/staging/#/builders/145/builds/2517

Differential Revision: https://reviews.llvm.org/D104518