[Zorg] Use ccache instead incremental build for openmp-offload-cuda-runtime.
Building with `LLVM_ENABLE_RUNTIMES=openmp` has the disadvantage that if clang changes, the runtime that was built with the previous clang is not rebuilt. That is, if the runtime is miscompiling due to a change in clang, this will only be detected in the next clean build.

For instance, commit rG1100e4aafea233bc8bbc307c5758a7d287ad3bae caused the libomptarget device runtime to miscompile, but the [[ https://lab.llvm.org/staging/#/builders/154/builds/1421 | openmp-offload-cuda-runtime builder ]] shows it as green. Tests only started failing with the [[ https://lab.llvm.org/staging/#/builders/154/builds/1427 | next clean build ]]. In production, this would have blamed the wrong commit.

In contrast, the openmp-offload-cuda-project builder started failing with the [[ https://lab.llvm.org/staging/#/builders/155/builds/1803 | expected commit ]].

Instead of building incrementally, use ccache to avoid this problem.

Reviewed By: gkistanova, tianshilei1992

Differential Revision: https://reviews.llvm.org/D106781