[zorg] Reorganize Polly builders.
Instead of a single build configuration, add 6 builders of the configurations (static, shared, shlib) x (LINK_POLLY_INTO_TOOLS=ON/OFF) for the 2 gce buildsbots. The test-suite build is extended to include external tests (SPEC benchmarks, povray) and moved to 2 other workers that should be able to compile it faster. Also, the verbosity is reduced (seems to have caused problem when the buildbot master receives too much data) and `-polly-process-unprofitable` is used for greater test coverage.

Differential Revision: https://reviews.llvm.org/D110048