Avoid building the entire tree and testing LLVM itself on MLIR builders
This reduces the amount of targets to build and tests. The build gets faster (see below) but
the main motivation right now is to not get notified by an mlir-* bot when there is a failure
in a LLVM test (there are enough other bots to cover these).

Build time gain is hard to evaluate because it is highly dependent on the machine.
Below are the results on my Linux machine (with the same cmake config as the
win-mlir-buildbot, so no python tests):
```
$ time ninja check-mlir -j 16
894.323 [0/1/2803] Running the MLIR regression tests

Testing Time: 98.80s
  Unsupported      : 163
  Passed           : 927
  Expectedly Failed:   1

real	16m41.103s
user	188m17.260s
sys	23m22.081s
```

After that, adding `ninja -j 16` results in an added:

```
$ time ninja -j 16
441.094 [0/1/767] Linking CXX executable bin/SpeculativeJIT

real	7m21.771s
user	86m25.291s
sys	15m5.505s
```

So it is ~1/3 of the build time that we're saving here.

Differential Revision: https://reviews.llvm.org/D110187