[builder][amdgpu] remove one builder - project, keep builder runtime
in preparation to move amdgpu openmp builder from Staging to buildbot,
i am retaining the runtime variant of the two builders we had in use during staging testing.
a second system is being ordered, and it will house a newer worker/builder that will be in staging.

Galina suggested that it was better to add new workers to a new machine if possible to avoid introducing stability issues in the buildbot.

Differential Revision: https://reviews.llvm.org/D111305