clang-ve-ninja: Use make targets as build steps
This turns 've-linux.py' into a driver to do annotated builds with
Makefiles, for now just 've-linux-steps.make'.  The 'get-steps' target
returns a list of make targets. The annotated builder creates a build
step for each of them.

The rationale for this is mostly the convenience of Makefiles and being
able to use the same script in buildbot and elsewhere.