[zorg] Remove the clang-ppc64le-linux bot
This patch aims to remove the clang-ppc64le-linux bot.

The clang-ppc64le-linux bot was originally meant to be a bot to build LLVM and
run LIT tests, in order to quickly notify of any regressions. However, building
on the bot is not particularly quick at all.

As this bot only performs a subset of the tests that the other bots runs, and
that the average build times shown on the webpage exceeds that of other bots
(such as the multistage bot), it defeats the purpose of having this "quick bot"
in the first place.

Differential Revision: https://reviews.llvm.org/D111813