Add BOLTBuilder: BOLT custom builder
The builder is based on UnifiedTreeBuilder.getCmakeWithNinjaBuildFactory, but adds one extra checkout step
between getLLVMBuildFactoryAndSourcecodeSteps and addCmakeSteps/addNinjaSteps.
Builder checks out binary tests from another repo, currently at https://github.com/rafaelauler/bolt-tests.
The plan is to transfer the repository over to llvm organization in the future.

The configuration is tested with a temporary build master, which points to the current BOLT repo instead of the monorepo.

This diff is a part of BOLT upstreaming:
https://lists.llvm.org/pipermail/llvm-dev/2021-December/154290.html

Differential revision: https://reviews.llvm.org/D116060