[zorg] Remove the clang-ppc64be-linux bot
This removes the clang-ppc64be-linux bot.

The clang-ppc64be-linux bot was originally meant to be a bot to build LLVM and
run LIT tests, in order to quickly notify of any regressions. However,
it does not have an average build time that is faster than other bots like the
clang-ppc64be-linux-lnt even though it is running a subset of this bots
test. It is therefore no longer useful as a "fast builder" for reporting
regression in the build or LIT testing and is being removed.

Reviewed By: gkistanova, amyk

Differential Revision: https://reviews.llvm.org/D117716