[clang-ppc64-aix] Set OBJECT_MODE as an builder wide env
Currently we pass this env to the build factory, but that means it only gets propogated to those factory steps which remember to pass the env into their respective shell commands.

On AIX, it's important that this env remain constant across the cmake/ninja/lit invocations or we will get into strange mixed bitmode situations, so set it as a builder env that will be inherited by all shell commands by default.

Reviewed By: ZarkoCA

Differential Revision: https://reviews.llvm.org/D118710