Limit test parallelism on Linaro's AArch64 compiler-rt bots
These run on a shared machine where the number of cores
is much greater than the number a single bot can expect
to use at once.

This is fine for compiling but for sensitive tests like
a few in compiler-rt, they can fail randomly depending on
how long they have to wait for resources.