[Buildbot] Updated configuration for Armv7/Aarch64 Win-x-Linux cross builders.
In order to updated the cmake cache file for the Linux-x-Win cross builds,
which has been changed to support building the libraries in the multitarget mode,
we need to update the builder configuration to handle running the library tests
properly. Currently we need to specify a proper target for the tests directly
within the lit test parameters (or ninja target).

Differential Revision: https://reviews.llvm.org/D120293