[llvm test suite] Treat NOEXE test result as a failure
Currently the SVE bots have been crashing when trying to compile
some of the bitcode tests. This was only noticed recently by chance
but it is not reported as a failure on the bots.

This makes it so that NOEXE counts as a failed test. This seems to
make sense because NOEXE is similar to unresolved tests.

I also spot checked other test-suite bots and found that none
of them had existing NOEXE results. (I thought there might be some
that are failing to build but this is expected to happen)

In any case, tests are guarded with cmake files to enable them
only if you have the right tools. So I think a failure to compile
should be a failure.

Differential Revision: https://reviews.llvm.org/D121282