[Fuchsia] Omit the list of projects and runtimes
The list of projects and runtimes is a part of Fuchsia-stage2.cmake
cache file now. This addresses the current failure which is due to a
mismatch between the configuration and the content of the cache file.

In addition, also include bolt and polly in the list of projects that
trigger our builders since bolt is already enabled in the Fuchsia CMake
cache file.

Differential Revision: https://reviews.llvm.org/D123282