Changing the ppc64[be|le] bots to TestSuiteBuilders
Update the ppc64 lnt bots to use `cmake` and `ninja` to build and run test-suite

To achieve this with minimal impact, most of the functionality from `UnifiedTreeBuilder.py`
is copied into `TestSuiteBuilder.py` instead of importing it so that the step naming convention
and ninja environment variables can be modified without impacting all the bots that
use `getCmakeBuildFactory()`.

Bots are renamed to reflect this change:
* clang-ppc64le-linux-lnt => clang-ppc64le-linux-test-suite
* clang-ppc64be-linux-lnt => clang-ppc64be-linux-test-suite

The clang-ppc64le-rhel bot is updated to run check-all instead of check on the llvm-lit testing step.

Reviewed By: gkistanova

Differential Revision: https://reviews.llvm.org/D128273