Raise timeout for Solaris builders
As described in D129837 <https://reviews.llvm.org/D129837>, the Solaris
builders (`solaris11-amd64` in particular) suffer from a `procfs` bug that
causes long runtimes for some sanitizer tests, which eventually get killed
after the default timeout of 1200 s.  As a workaround, this patch increases
the timeout to allow them to finish.

Differential Revision: https://reviews.llvm.org/D129961