[zorg] Run test-suite with flang-new
At the moment, `clang-aarch64-full-2stage` is running the
`llvm-test-suite` with `flang-to-external-fc`, which is a wrapper script
involving both flang and gfortran.

This patch switches to using the `flang-new` driver, therefore removing
our dependence on gfortran.

However, support for generating executables with `flang-new` is still
experimental and needs a few flags to help it along.

Furthermore, we need an environment variable when running the
test-suite (see
https://flang.llvm.org/docs/FortranLLVMTestSuite.html#:~:text=to%20set%20the-,NO_STOP_MESSAGE,-environment%20variable%20to).

Differential Revision: https://reviews.llvm.org/D130628