[mlgo] Enable using TFLite instead of full Tensorflow C API lib
For the training case, TFLite is more appropriate: it plays well with
cmake (thanks to @etrhosek's work to making that turnkey); can be
statically linked; and it's faster.

This patch introduces the support side-by-side the support for the full
TF C API library; once we land the ml-compiler-opt and LLVM changes and
make sure the build bots are still green, we will remove the full TF C
API support.

Differential Revision: https://reviews.llvm.org/D130923