Fix AArch64 SVE bots
The SVE bots are running the llvm-test-suite with flang, but they're not
setting NO_STOP_MESSAGE in the environment. This causes most of the
tests to fail, as expected (see
https://flang.llvm.org/docs/FortranLLVMTestSuite.html#:~:text=to%20set%20the-,NO_STOP_MESSAGE,-environment%20variable%20to)

This patch makes sure to set NO_STOP_MESSAGE to 1 for all the SVE bots.
As far as I can tell, no other bots run the test-suite with flang.

Differential Revision: https://reviews.llvm.org/D132351