[lldb-cmake] Turn Debug Info test failures into pipeline errors
Currently the "Debug Info" tests fail but that is not reflected
in the pipeline. We do a `set +e` before we run the tests to
overide the default Jenkins shell's `set -e`; this hides the
test failures. Remove this for now and see if fixing the test
failures will be enough to make this pipeline stage consistent
with the behaviour of the other stages.

Differential Revision: https://reviews.llvm.org/D132391