[lldb-cmake] Turn Debug Info test failures into pipeline errors
Currently the "Debug Info" tests fail but that is not reflected
in the pipeline. We do a `set +e` before we run the tests to
overide the default Jenkins shell's `set -e`; this hides the
test failures. We wrap the `monobuild_repo.py` invocation
in a `catchError` so the subsequent stages in the pipeline
still execute (in particular the `Test (LLDB)` build).

Differential Revision: https://reviews.llvm.org/D132420