[Windows on Arm] Don't use an env var for the msvc compatibility version
Long story short the cmake command isn't run in a shell but ninja
generated commands are. This is why %MSVC_VER% wasn't expanded but
%BUILTINS% wasn't used at configure time so it didn't matter there.

I looked at setting LLVM_TEMPORARILY_ALLOW_OLD_TOOLCHAIN but
both the minimum and soft error versions are set to 19.27
so it doesn't help us.

Instead just hardcode the expected verison. We'll probably always
be behind as we're using a released clang-cl to build latest
clang-cl but at least it won't change that often.

Differential Revision: https://reviews.llvm.org/D132651