Change PowerPC64 bots to build runtimes with -DLLVM_ENABLE_RUNTIMES
All of the powerpc 64-bit little and big endian buildbots build their
runtimes with -DLLVM_ENABLE_PROJECTS in the cmake command; this method
is deprecated.
This patch updates all of the ppc64 bots, except the sanitizer bot,
to build runtimes with -DLLVM_ENABLE_RUNTIMES and adds a check-runtimes
test step as check-all does not currently check runtimes.

Reviewed By: amyk

Differential Revision: https://reviews.llvm.org/D133940