[libc++] Remove disabled AIX bot for libc++
This patch removes the last libc++ builder on BuildBot, which has been
disabled ever since I've noticed it. We have a builder for AIX on
BuildKite, so I assume this had simply been forgotten.

It also removes the LibcxxAndAbiBuilder factory, which has been unusued
and unmaintained since we've moved to BuildKite. We could also keep it
around, however IMO it is better to remove all traces of libc++ CI in
BuildBot to avoid confusion, since we don't use it anymore and it is
not the preferred way of testing that project.

Differential Revision: https://reviews.llvm.org/D134359