Add directory and readme for premerge tests.
Previously premerge infrastructure code was hosted at
https://github.com/google/llvm-premerge-checks as the infrastructure was
developed and maintained by Google developers initially. On the other
hand it was not a best option for many contributors due to the need to
sign Google's CLA; it was also meant to be a community-owned project.

The plan is to move / rewrite most of the code from
https://github.com/google/llvm-premerge-checks here in zorg under LLVM
license.