[sanitizers] Controlling the build width and targets for PowerPC sanitizers
The cmake sanitizer buildbots use the `nproc` value for the build
width by default, this means the PowerPC sanitizer buildbots were
building and testing at the max parallelism their host machines could
handle.
This patch adds a build and test width specification for the PPC
sanitizers and sets the build targets to only build for PowerPC
with the intension of reducing the load the sanitizers have on their
host while improving build times. It fixes up cmake argument passing
to facilitate the change.

Reviewed By: amyk

Differential Revision: https://reviews.llvm.org/D135806