[builder][standalone-build] Have separate trace/verbose vars to control the build output
In D138114 I've introduced `CMAKE_TRACE` and mistakenly assumed it can be applied to `cmake --build` and `cmake --install` calls as well. That isn't the case. There only exist `--verbose` for build and install commands in cmake. That is why I have removed `CMAKE_TRACE` and replaced it with multiple variables to control `llvm`'s, `clang`'s, and `lld`'s configure, build, check and install steps.

Differential Revision: https://reviews.llvm.org/D138120