Use Debian unstable for C++20 buildbot
We want to use libstdc++-12 as it adds things that cause breakages with
the current release of Clang, e.g. constexpr to vector members.
Debian stable does not ship with libstdc++-12 and apt.llvm.org does not
have a build for Debian Testing. So Debian unstable is the only viable
option for this buildbot.