Generate ENABLE_RUNTIMES in ClangBuilder
In ca793509f5a9b26395400dcb92b344b1e055e949 I passed down
enable_runtimes to ClangBuilder but did not realise that
it was not LLVMBuildFactory that generates the CMake arguments.

LLVMBuildFactory simply holds properties of the build. In this
case what projects and runtimes we rely on.

ClangBuilder is the one making the CMake arguments, so I've updated
that to emit the right options this time instead of putting everything
in PROJECTS.

I've adjusted two of the PPC bots that were manually putting compiler-rt
into RUNTIMES.

Differential Revision: https://reviews.llvm.org/D144858