[s390x] Put compiler-rt in runtimes
This moves compiler-rt from ENABLE_PROJECTS, to ENABLE_RUNTIMES
for these bots:
* clang-s390x-linux
* clang-s390x-linux-multistage
* clang-s390x-linux-lnt

This matches the recommended way to build compiler-rt.

getClangCMakeBuildFactory's current default is enable_runtimes=None
which means the runtimes build is not used. "auto" puts anything known
to be a runtimes project into the list of runtimes instead of projects.

Differential Revision: https://reviews.llvm.org/D146505