[solaris] Explicitly opt out of the runtimes build
As described in https://github.com/llvm/llvm-project/issues/54196
and https://github.com/llvm/llvm-project/issues/57781, there are
ongoing issues with the runtimes build for Solaris and others.

This patch does not change what these bots do, just makes it explicit
so that when the default switches in future they are not effected.

Differential Revision: https://reviews.llvm.org/D146521