[m68k] Change compiler-rt settings
For clang-m68k-linux-cross it appears that compiler-rt is
not needed. It does not build it for m68k and there are other
bots covering compiler-rt on x86 (which hosts the build).

I was unable to confirm the configuration of clang-m68k-linux
(https://lab.llvm.org/buildbot/#/builders/159 doesn't show any builds).
And cannot search for it on staging at the moment.

So I am opting it out of the runtimes build for compiler-rt
just to be on the safe side.

Differential Revision: https://reviews.llvm.org/D146532