[OpenMP] Remove unnecceary OMPT flag from builder factory
The OMPT option defaults to "ON" in the OpenMP project CMake. To remove
potential confusion whether OMPT support is enabled in a buildbot,
remove this argument. It was only able to *enable* OMPT support and is
now useless.

If OMPT support should be disabled in a buildbot, this should be made
explicit via the additional CMake flag.

Differential Revision: https://reviews.llvm.org/D146703