Remove llvm-linux-avr builder + worker
This bot has been red for over 14 weeks. I reached out to the bot
owner via email and the review for these changes, but did not hear back
from them. So removing the bot as being stale (we can bring it back
online easily enough if the bot owner needs it again).

Differential Revision: https://reviews.llvm.org/D147702