[ClangBuilder] Change default of enable_runtimes to "auto"
Prior to this I went through and made all bots explicitly opt
into, or out of, the runtimes build.

Now that's done the majority are opting in, so make that the default
and remove the explicit opt ins.

"auto" means that we automatically put any runtimes projects
(libcxx, compiler-rt etc.) in LLVM_ENABLE_RUNTIMES instead of
LLVM_ENABLE_PROJECTS.