[OpenMP][LibC] Run libc GPU tests with 1 thread
Executing the libc GPU tests multithreaded may overload the GPU runtimes
for CUDA or HSA, which results in test errors. In addition, libc seems
to not respect the `-j` option of ninja, when it is invoked at the top
level.

This patch adds special handling of the `check-libc` rule when passed as
additional check to the OpenMP builder. Ninja will thus first `cd` into
the runtimes build directory to make the libc testing respect the given
amount of parallelism.
We then limit the number of threads to `1` as to minimize the potential
for any errors with GPU runtimes.

Differential Revision: https://reviews.llvm.org/D149951