[Clang] Enable Clang Sphinx man page build
This patch makes the current clang-sphinx-docs also build the Clang man
pages. There have been some breakages within this functionality that
haven't been caught for a while (see https://reviews.llvm.org/D149809).
Having coverage of this would probably be a good idea as it shouldn't
cost that much and ensure that this build configuration works.

However, the build was also broken for a fairly significant amount of
time previously and no issues were filed other than the fix I opened, so
maybe this configuration is not particularly important to consumers.

Reviewed By: gkistanova

Differential Revision: https://reviews.llvm.org/D150924