Adding a clobber test suite step to the TestSuiteBuilder
The TestSuiteBuilder is suppose to build and test the
llvm-test-suite with each new compiler it builds but does
not clean the test-suite build directory on each build request.
This patch adds a clean step to remove the test-suite build
directory on each build request allowing each newly built compiler
to test test-suite thoroughly.

Reviewed By: amyk

Differential Revision: https://reviews.llvm.org/D151656