[Arm] Remove native Armv7-a builders/workers
clang-armv7-vfpv3-full-2stage and clang-thumbv7-full-2stage have
been unstable for a very long time and have never been that useful
due to a long cycle time.

We (Linaro) plan to bring these back in the future in a more useful
form. In the meantime, we do also have 32 bit and Armv7 builders that
run on Armv8-a machines.

Differential Revision: https://reviews.llvm.org/D153543