[ARM] Remove clang-native-arm-lnt-perf
Although this bot has been green 99% of the time, it is not providing
us anything that a the faster clang-armv7-lnt bot isn't doing already.

Aside from the fact that clang-armv7-lnt actually runs on an armv8-a
machine, but we (Linaro) have a plan to make up for that if it ever
becomes an issue.

Differential Revision: https://reviews.llvm.org/D154483