[flang] Add builder/worker for debug build with reverse iteration
For https://discourse.llvm.org/t/reverse-iteration-bots/72224.

This does not warrant a whole new builder, so I'm adding reverse
iteration to the debug configuration.

I thought of a few other places but:
* The release builder tests what we will later ship, so I don't
  want to change the configuration.
* Release with asserts is a very common developer configuration
  so it helps to know if problems occur with that only.
* The rest are build types, out of tree, shared libs, etc. which
  are niche enough I don't want reversed iterator problems polluting
  them.

Initially I'm not removing the existing worker, so that the two
can run in parallel with the reverse iterator bot on silent.

Once it's been setup and is stable, it'll move to normal and
I'll remove the old worker and builder.

Differential Revision: https://reviews.llvm.org/D157274