[AArch64] Enable MLIR SVE integration tests on 2 stage SVE bots
Previously these ran only on AArch64 2 stage (using QEMU), and the VLA/VLS
single stage SVE bots.

Later we added/brought back the 2 stage SVE bots so let's enable the
testing there too. We recently had them failing everywhere but 2 stage
SVE and the inconsistency made triaging it more confusing than it had
to be.

The 2 stage bots share the same hardware and containers with the single
stage so there should not be any new issues that aren't real problems.

Differential Revision: https://reviews.llvm.org/D158092