[AArch64][Flang] Add new flang-aarch64-libxx builder, remove flang-aarch64-latest-clang
This renames flang-aarch64-latest-clang to reflect what it actually
does. It does not use the latest clang release and it does not build
a clang as the first step. This is what we have 2 stage builders for.

As flang is a big user of newer C++ features, I do think this bot is
still useful so I'm renaming it to make that obvious.

This change adds the new worker/builder and removes the old ones.
So what I will do prior to landing this is remove the old workers
on our side so there are no lingering connection attempts.

The new worker will start life on silent as per the usual process.

Differential Revision: https://reviews.llvm.org/D159102