[RISCV] Add new riscv32 buildbot
This new bot will copy and run all the tests in qemu-system-riscv32,
with an image built with all the libc dependencies to run the tests.

We are using the CROSSCOMPILING flag available on cmake to pass the
script that will handle the scp part and execute the tests via ssh
remote commands.

This patch assumes that the qemu image is already running.

Differential Revision: https://reviews.llvm.org/D158284