Production build bots to report build statuses to github.
Since we moved to github PRs and people are building their workflows more github centric, I make all the production build bots (except any with the tag "silent") reporting build statuses to github.

We have this for few selected builders for years. Now is the time to make it for all.