[LibC++][PowerPC] Enable libcxx testing on clang-ppc64le-rhel (#62)
This patch enables the build and test of the libcxx and libcxxabi
runtimes on the clang-ppc64le-rhel buildbot. The compiler toolchain
is updated to use gcc-toolset-12 in order to resolve some libcxx lit
test case failures, we no longer specify the `LLVM_BINUTILS_INCDIR`
since the gold plugin is not being build and ccache is turned on to
improve build times.