Accept new requests on SVE builders only if idle
Use nextBuild on SVE builders to accept new build requests only if
the corresponding builder is not currently building anything.

This is needed on SVE bots because any of the 4 available workers
can accept requests from any of the 4 SVE builders, and sometimes
some builders use multiple workers while others end up with none
available. This is aggravated by the fact that the same builders
may use more than one workers for several times in a row, causing
other builders to starve (clang-aarch64-sve-vls, for instance, did
not build anything in the last 3 days).

The buildbot documentation (2.5.2.8. Prioritizing Builders) says
that builds are started in the builder with the oldest pending
request, but it seems this is not working, possibly because the
collapse requests feature ends up resetting submittedAt time.