Update the builder configurations on as-builder-4(-rel) worker hosts. (#84)
* reduce a build path for the builders to avoid failures for the unix
socket related tests (the expensive check builders).
* remove TMP/TEMP substitution for the release builders (we don't use a
ramdisk for these builders).

We start getting the failures for the unix socket related tests when
overriding TMP/TEMP env variables with the longer path on the Linux
hosts, The tests create an unique path for the sockets based on TMP +
uniqie name + socket name. The test gets failed when a total length of
this path overflows a limit in 108 characters.