Set jobs=None for Linaro LLDB builders
Turns out the LLDB builder class defaults to `jobs="%(jobs)s"`
which expects every worker to have a `jobs` property. The other
classes look for it but handle it not being there.

There is a proper fix to be done here but it needs review so for now,
set Linaro's builders to `jobs=None` as a workaround.