[LLDB] Handle missing "jobs" property on workers
Previously the builder class argument "jobs" defaulted to a string to use
with `WithProperties`. This assumed that every worker has
a "jobs" property, which is no longer true after changes
to Linaro's workers.

The behaviour of `WithProperties` is detailed here:
https://github.com/buildbot/buildbot/blob/6e7218792bc324584af980bd1abb4d8b6a5055b8/master/docs/manual/configuration/properties.rst#withproperties

This changes the argument to the builder class to be
`jobs=None` by default. Keeping the previous behaviour
but avoiding a KeyError later.

If a value is passed here it'll be added as `-j<value>`,
otherwise we use the worker's "jobs" property as before.
Except that we now account for the fact that it might be missing
(using code adapted from `NinjaCommand`).

So the behaviour is the same but we won't get a key error
for Linaro's bots that don't have "jobs".

The lldb x86 bot does have "jobs" and it will be passing
`jobs=None` to the builder class, which makes it use
the "jobs" property as it did before.

According to Buildbot docs, `WithProperties` should be replaced
with `Iterpolate`, but I think doing so here would only complicate
this fix.