[LLDB] Remove builder class jobs option, only use worker's jobs property if present (#102)
Previously the builder class argument "jobs" defaulted to a string to
use with `WithProperties`. This assumed that every worker has a "jobs"
property, which is no longer true after changes to Linaro's workers.

The behaviour of `WithProperties` is detailed here:
https://github.com/buildbot/buildbot/blob/6e7218792bc324584af980bd1abb4d8b6a5055b8/master/docs/manual/configuration/properties.rst#withproperties

As no builder was using the `jobs=` argument I have removed it and
adapted code from `NinjaCommand` to add `-j<jobs>` only if the property
exists.

Meaning that Linaro's workers that do not have a "jobs" property will
not be limited. `lldb-x86_64-debian` will have a `-j` because its worker
specifies `'jobs': 72`.

According to Buildbot docs, `WithProperties` should be replaced with
`Interpolate`, but I think doing so here would only complicate this
change.