[OpenMP] Disabling SDMA transfers AMDGPU offload bots
This disables host/device and device/host transfers via the SDMA engines
for the OpenMP Offload AMDGPU buildbots.
The motivation is a known issue of a potentially missed synchronization
tracked via https://github.com/ROCm/ROCm/issues/2616.

This change was rolled-out to one of the staging bots previously via
https://github.com/llvm/llvm-zorg/pull/114 and seemed to remove the
flakiness we were observing.