[lldb] Use NinjaCommand instead of manually building options (#122)
Which is what I tried to do in 330340b14f6325d76c96cb1713d29c4258e541b8,
but forgot one detail. If there is no property the options have empty
string added to them.

This means you get this:
`ninja "" ""`

And ninja thinks that those are targets to build, which don't exist of
course.

NinjaCommand already handles this by removing any empty arguments as the
final step of building the options. So use that for our ninja steps
instead.