[libc] Update libc build config
How to build libc for GPU is about to change significantly, which
requires an update to the way it is built in the bot.

After some back and forth, it appears that it cannot be easily added to
the -DLLVM_ENABLE_RUNTIMES and instead requires some special handling.
Thus, for the libc bot we intentionally work around the
enable_runtimes=auto option and instead enable 'openmp' and
'compiler-rt' manually in the bot specification.
In addition, we specify the new CMake flags for the libc build on
AMDGPU.