Add -DLLVM_DISABLE_ASSEMBLY_FILES=ON to work around a bug in Apple Clang (#129)
Two build-id related LLD tests are failing on the MacOS builder I am
trying to bring up apparently due to a bug in the version of Apple clang
installed on the machine. (See llvm/llvm-project#83940 and
llvm/llvm-project#81967).

The suggested workaround of adding `-DLLVM_DISABLE_ASSEMBLY_FILES=ON`
seems to work on the bot, so add that to the configuration.