[OpenMP/Offload] Enable Offlaod runtime and add check
This enables the new "offload" runtime that
https://github.com/llvm/llvm-project/pull/75125 will create. In
addition, it adds a new lit test target.
I removed the extra CMake argument that would again specify which
runtimes to build manually, so that we rely on the factory to add the
respective runtimes.

Since the PR does not make `check-libomptarget` available in the
top level configuration, we have to change the ninja working directory
into the runtimes build folder, where the target is known. Whether or
not this works with this crude "additional lit checks" mechanism is an
experiment.